# server/docs/filehandler-openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "File Sharing API",
    "description" : "API for File Sharing Application - Upload, Download, Share files",
    "contact" : {
      "name" : "File Sharing Team",
      "email" : "support@filesharing.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5000",
    "description" : "Local Development Server"
  }, {
    "url" : "https://api.filesharing.com",
    "description" : "Production Server"
  } ],
  "security" : [ {
    "Bearer Authentication" : [ ]
  } ],
  "paths" : {
...
    "/api/audit-logs/asset/{assetId}" : {
      "get" : {
        "tags" : [ "audit-log-controller" ],
        "operationId" : "getAssetAuditLogs",
        "parameters" : [ {
          "name" : "assetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "versionNumber",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponsePageResultAuditLogItemDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AuditLogItemDTO" : {
        "type" : "object",
        "properties" : {
          "iconType" : {
            "type" : "string"
          },
          "action" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "timeAgo" : {
            "type" : "string"
          },
          "actorName" : {
            "type" : "string"
          }
        }
      },
      "CommonResponsePageResultAuditLogItemDTO" : {
        "type" : "object",
        "properties" : {
          "isSuccessful" : {
            "type" : "boolean"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResultAuditLogItemDTO"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "PageResultAuditLogItemDTO" : {
        "type" : "object",
        "properties" : {
          "totalCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditLogItemDTO"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "type" : "http",
        "description" : "Enter JWT token",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}